// shared/components/shared/checkout-components/checkout-form-schema.ts
import {z} from "zod";

export const checkoutFormSchema = z.object({
    firstName: z.string().min(2, {message: "Имя должно содержать не менее 2 символов"}),
    lastName: z.string().min(2, {message: "Фамилия должна содержать не менее 2 символов"}),
    email: z.string().email({message: "Некорректная почта"}),
    phone: z.string().min(10, {message: "Некорректный телефон"}),
    address: z.string().min(5, {message: "Адрес должен содержать не менее 5 символов"}),
    comment: z.string().optional()
});

export type CheckoutFormValues = z.infer<typeof checkoutFormSchema>;